+++ CHANGELOG.md
@@ -16,11 +16,133 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-## 2019-07-09
+## 2019-07-13
 
 ### Added [Set-RubrikSLA]
 
 * Added new `Set-RubrikSLA` cmdlet to update an existing SLA Domain. This addresses issue [283](https://github.com/rubrikinc/rubrik-sdk-for-powershell/issues/283)
+
+## 2019-07-12
+
+### Fixed issues with parsing of complex JSON payloads
+
+* Get-RubrikReportData now correctly returns all data
+* Performance of JSON parsing improved for PowerShell 6 and later
+
+## 2019-07-11
+
+### Added [Various Unit Tests]
+
+* Added unit tests for Get-RubrikManagedVolume, Get-RubrikManagedVolumeExport, Get-RubrikVolumeGroup, Get-RubrikVolumeGroupMount, New-RubrikManagedVolume, New-RubrikManagedVolumeExport, New-RubrikVolumeGroupMount, Remove-RubrikManagedVolume, Remove-RubrikManagedVolumeExport, Remove-RubrikVolumeGroupMount, Set-RubrikManagedVolume, Start-RubrikManagedVolumeSnapshot, Start-RubrikManagedVolumeSnapshot to address [Issue 340](https://github.com/rubrikinc/rubrik-sdk-for-powershell/issues/340)
+* Note - Set-RubrikVolumeFilterDriver already contained a unit test.
+
+## 2019-07-10
+
+### Fixed [Disconnect-Rubrik & Connect-Rubrik]
+
+* Added global attribute 'authType' to $rubrikconnection to remove reliance on userId.
+* Added logic to disconnect to address the deletion of tokens when using token based authenticaion as per [Issue 363](https://github.com/rubrikinc/rubrik-sdk-for-powershell/issues/363)
+
+## 2019-07-08
+
+### Added [Unit Tests]
+
+* Added unit tests for Update-RubrikHost and Update-RubrikvCenter to address [Issue 339](https://github.com/rubrikinc/rubrik-sdk-for-powershell/issues/339)
+
+### Changed [Get-RubrikDatabase]
+
+* Added ability to specify -DetailedObject on Get-RubrikDatabase to address [Issue 354](https://github.com/rubrikinc/rubrik-sdk-for-powershell/issues/354)
+
+## 2019-07-03
+
+### Added [Unit tests]
+
+* Added unit tests for Get-RubrikHyperVVM, Get-RubrikNutanixVM, Move-RubrikMountVMDK, New-RubrikVMDKMount, Protect-RubrikHyperVVM ,Protect-RubrikNutanixVM. Protect-RubrikVM, Set-RubrikHyperVVM, Set-RubrikNutanixVM, Set-RubrikVM
+
+### Changed [Fixes for unit tests]
+
+* Minor updates to parameter configurations of Get-RubrikHyperVVM, Get-RubrikNutanixVM, Move-RubrikMountVMDK, New-RubrikVMDKMount, Protect-RubrikHyperVVM ,Protect-RubrikNutanixVM. Protect-RubrikVM, Set-RubrikHyperVVM, Set-RubrikNutanixVM, Set-RubrikVM so they pass associated unit tests
+
+### Changed [Get-RubrikDatabase] - Relic parameter
+
+* Parameter now has 3 states -Relic -Relic:$false or not specified
+
+## 2019-07-02
+
+### Added Unit Tests [Export-RubrikReport & Export-RubrikDatabase]
+
+* Added Unit Tests for Export-RubrikReport & Export-RubrikDatabase to address [Issue 333](https://github.com/rubrikinc/rubrik-sdk-for-powershell/issues/333)
+
+## 2019-06-29
+
+### Changed Standardized usage of Out-Null over the module
+
+* To improve PowerShell 6, and onwards compatibility we have standardized on using | Out-Null
+
+### Added [New-RubrikSnapshot] Additional help examples in
+
+* Added examples of how to do Full backups of Oracle and MSSQL databases
+
+## 2019-06-28
+
+### Changed [quick-start.md] - Additional download instructions
+
+* Added a 4th option for downloading and distributing the Rubrik SDK for PowerShell
+
+## 2019-06-27
+
+### Changed [Sync-RubrikAnnotation]
+
+* Added -DetailedObject to Get-RubrikVM in order to return the snapshots
+* Added a third annotation to store the date of the latest Rubrik snapshot.
+* Added associated unit tests for Sync-RubrikAnnotation
+
+### Changed [New-RubrikSnapshot]
+
+* Cmdlet will now display a warning if -ForceFull is set on any other protected object other than Oracle or SQL databases.
+* This is just a warning and the cmdlet will continue to run, performing an incremental backup.
+* This addresses [315](https://github.com/rubrikinc/rubrik-sdk-for-powershell/issues/315)
+
+### Fixed [Protect-RubrikTag]
+
+* modified Protect-RubrikTag in order to ignore relic's when retrieving the vCenter UUID.
+* Addresses [Issue 311](https://github.com/rubrikinc/rubrik-sdk-for-powershell/issues/311)
+* added associated Unit test for the cmdlet.
+
+## 2019-06-26
+
+### Added [Tests for Get-RubrikHost]
+
+* Added unit test for Get-RubrikHost cmdlet
+
+### Added [Get-RubrikAPIToken cmdlet]
+
+* Added Get-RubrikAPIToken cmdlet to address [321](https://github.com/rubrikinc/rubrik-sdk-for-powershell/issues/321) and associated unit test.
+
+### Modified [New-RubrikSnapshot cmdlet]
+
+* Added support for Oracle to New-RubrikSnapshot
+* Added tests for New-RubrikSnapshot
+
+### Fixed [Get-RubrikHost, Get-RubrikVM, Get-RubrikOracleDB]
+
+* Added formating around $result to convert to an array in order to support -DetailedObject with older versions of Powershell.  Addresses [319](https://github.com/rubrikinc/rubrik-sdk-for-powershell/issues/319)
+
+## 2019-06-25
+
+### Added [New New-RubrikAPIToken cmdlet]
+
+* Added New-RubrikAPIToken cmdlet to address [316](https://github.com/rubrikinc/rubrik-sdk-for-powershell/issues/316) and associated unit test.
+
+### Added [New Remove-RubrikAPIToken cmdlet]
+
+* Added Remove-RubrikAPIToken cmdlet to address [316](https://github.com/rubrikinc/rubrik-sdk-for-powershell/issues/316) and associated unit test.
+
+## 2019-06-24
+
+### Added [New Get-RubrikOracleDB cmdlet]
+
+* Added Get-RubrikOracleDB cmdlet to address [255](https://github.com/rubrikinc/rubrik-sdk-for-powershell/issues/255) and associated unit test
 
 ## 2019-06-20
 
@@ -64,7 +186,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added [New-RubrikManagedVolume update]
 
 * Added `-ApplicationTag` parameter support to New-RubrikManagedVolume so users can specify which application the managed volume will be used for. This addresses issue [285](https://github.com/rubrikinc/rubrik-sdk-for-powershell/issues/285).
-
 
 ## 2019-05-30